+++ global.d.ts
@@ -1,0 +1,3 @@
+declare interface AqaPackageSection {
+    verbose?: boolean
+}